requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
google-generativeai==0.5.4
protobuf==3.20.3
grpcio==1.60.0